testTools merge controlClass
